修复部分警告，将部分%ld修改为%zi作为字符的打印信息

+++ AxcUIKit/SampleVC/AxcActivityHUDVC.m
@@ -152,7 +152,7 @@ UICollectionViewDelegateFlowLayout
                                                                                          forIndexPath:indexPath];
     cell.backgroundColor = [UIColor AxcUI_ArcColor]; // 预设随机色
 
-    cell.iconImageView.image = [UIImage imageNamed:[NSString stringWithFormat:@"HUD_%ld",indexPath.row]];
+    cell.iconImageView.image = [UIImage imageNamed:[NSString stringWithFormat:@"HUD_%ld",(long)indexPath.row]];
     
     return cell;
 }
